Ansible 9 Loops
This lesson examines the use of loops in Ansible to perform the same type of tasks. Loops allow you to simplify playbooks, especially when you need to perform many similar actions.
Sample code:
- hosts: all
  tasks:
    - name: install packages
      ansible.builtin.apt:
        name: "{{ item }}"
        state: latest
      loop:
        - redis-server
        - postgresql
      become: yes
This example uses the apt module to install the redis-server and postgresql packages. The loop loop allows you to perform a task sequentially for each item in the list. The {{ item }} variable is used to substitute the current list item in the name parameter.
Loops can be used with any Ansible modules. In the example above, the loop applies to only one task, but it can also be used for multiple tasks within the same playbook.
Result:
Loops in Ansible make it easier to perform the same type of tasks.
Loops can be used with any Ansible modules.
The {{ item }} variable is used to substitute the current list item in the module parameters.
Loops can be used to perform multiple tasks within a single playbook.