Ansible 11 Filters
In this lesson, we learned the basics of using filters in Ansible. Filters allow you to transform data and manage it in the process of completing tasks. They are applied to variables or values inside double curly braces {{ }} using the vertical line symbol |.
We have looked at several examples of filters, such as min, max, unique and random, which allow you to find the minimum and maximum values, unique values and select a random value from a list.
We also looked at examples of other popular filters, such as ipaddr, hash, basename, dirname and expanduser, which allow you to work with IP addresses, create hashes, extract the file name and directory path, as well as disclose paths.
Filters in Ansible provide ample opportunities for processing data and converting it to the desired format. They can be useful when choosing a server according to certain parameters, processing JSON structures, replacing text using regular expressions, and much more.
In addition, you can create your own Python filters in Ansible, which allows you to expand the capabilities of the tool and adapt it to specific tasks.