Ansible 12 Registration of the result
This lesson discusses the mechanism for registering the results of tasks in Ansible. This mechanism allows you to save the result of one task and use it in other tasks.
In the example above, a task is first performed using the shell module, which outputs a list of the user's home directories. The result of this task is saved to the home_dirs variable using the register key.
Then, using the cron module, the home directories are added to the list of cron tasks. To do this, the with_items loop is used, which runs through each line of output from the home_dirs variable.
At the end of the playbook, using the debug module, the contents of the home_dirs variable are output to make sure that the results are saved correctly.
Bottom line:
The results registration mechanism allows you to save the result of the task to a variable.
The result can be used in other tasks.
The results are registered using the register key.
The result is saved as a hash containing information about the completion of the task.
The result can be used in a loop using the with_items key.