Ansible 13 Roles
Ansible is an automation tool that allows you to manage the configuration and deployment of systems. It is based on the YAML language and uses the SSH node management model.
The main idea of Ansible is to describe the desired state of the system in the form of playbooks that contain a set of tasks. Tasks can be various operations, such as installing packages, configuring configuration files, running commands, etc.
Ansible also supports the use of roles, which are sets of tasks and variables designed to perform certain functions or install certain components. Roles can be created and shared with other users via Ansible Galaxy, which allows you to reuse ready-made solutions.
So, the basic concepts and capabilities of Ansible:
Playbooks: describe the desired state of the system and contain a set of tasks to achieve it.
Tasks: Operations to be performed on nodes.
Roles: Sets of tasks and variables designed to perform certain functions or install certain components.
Ansible Galaxy: A shared directory of roles available for use and sharing with other users.
Variables: used to customize the behavior of playbooks and roles.
Modules: Provide functionality to perform tasks such as managing packages, files, services, etc
. SSH: Ansible uses SSH to manage nodes, so it requires an SSH connection and proper access settings.
As a result, Ansible automates repetitive tasks, simplifies system configuration and deployment management, and facilitates collaboration and reuse of off-the-shelf solutions through Ansible Galaxy.