Ansible 14 Collections
The lesson explains that collections in Ansible are a distribution format for a linked set of playbooks, roles, modules, and plugins. They were created to simplify the support of a large number of modules that were previously built into Ansible. All built-in modules have been migrated to Ansible Collections, which can be installed as roles via Ansible Galaxy.
Ansible Collections are stored on GitHub (https://github.com/ansible-collections ) and include various modules, such as modules for working with cloud providers, databases, network devices and other technologies.
Collections are named according to the "namespace.collection.module" principle. For example, the collection for working with the PostgreSQL database is named "community.postgresql.postgresql_db".
Collections can be installed using the "ansible-galaxy collection install" command, and the list of installed collections can be viewed using the "ansible-galaxy collection list" command.
Collections allow you to use ready-made modules and roles to automate various tasks such as database management, configuring network devices, installing and configuring applications, and much more. They greatly simplify the development and support of playbooks and roles in Ansible.
Result:
Collections in Ansible are a distribution format for a linked set of playbooks, roles, modules, and plugins.
They were created to simplify the support of a large number of modules that were previously built into Ansible.
The collections are stored on GitHub and can be installed using the "ansible-galaxy collection install" command.
Collections allow you to use ready-made modules and roles to automate various tasks.