Ansible 15 Ansible Vault
Ansible storage is a mechanism for encrypting variables and files in Ansible. It allows you to hide secret data, such as passwords or keys, from unauthorized persons.
For use in Ansible Vault, it may have been created from scratch. This can be done using the ansible encryption store command, specifying the path to the file to be encrypted. When you run this command, you will be prompted for a password that you need to remember or save in a safe place.
Encrypted into a file, you can use Ansible playbooks by specifying the option --ASK-storage-transfer when starting the playbook. Ansible locks up steam, distributes data, and fills in unsolvable tasks. The file will remain encrypted.
For convenience, you can specify the path to the password file using the password -vault-password-file. In this case, the password will be taken from the specified file.
To edit the user interface, you can use the ansible-vault edit command, which is edited in the editor for silence. In this case, the file will remain encrypted.
Using Ansible Vault allows you to access confidential data from accidental access to private access, ensuring security when working with Ansible.