JS: HTTP Server 1 Server Implementation
server.js
Implement an http server, which is the interface for accessing the telephone directory.
The directory is represented by a text file phonebook.txt . Its format:
1 | Carleton Sporer | 197.328.3450
2 | Ashton Bogisich | 244.742.7016
3 | Valentin Auer | 964.685.7490
4 | Buddy Kuvalis | 356.157.9872
Each line represents a separate record about a person. It contains three values separated by the | symbol. 1 is the id, 2 is the name, 3 is the phone.
When requesting to / the server should return the following body:
Welcome to The Phonebook
Records count: <the number of lines in the phonebook file.txt>
Flow
In addition to running tests, be sure to try to "play" with the written server. To do this, use the make start command, which starts your server. After that, you can make requests to it using telnet or, for example, curl: curl localhost:8080. Don't forget to restart the server when the source code changes. To stop the server, use the ctrl+c combination.