JS: HTTP Server 2 Request
index.js
Implement the logic of parsing the phonebook file.txt to the following format:
{
  <id>: {
    name: <name>,
    phone: <phone>,
  },
  <id>: {
    name: <name>,
    phone: <phone>,
  },
}
Where <id> is the identifier of a specific record, and <name> and <phone> are the name and phone number of the corresponding record.
server.js
The server allows you to make search queries for all records that match the search criteria. The criterion is the part of the first /last name by which the comparison is made.
In a request to the server via the link /search, one parameter must be passed: q, the value of which will be a substring. If matching is found, the server returns the data in the following format (for the miss substring):
Miss Arlo Barrows, 328-949-3924
Miss Bernadette Conn, 249.059.5515
Miss Savannah Dicki, 157.463.3368
Miss Rudy Brown, 779-703-0150`
Please note that the case is not taken into account when matching.
If the request to the server is executed without parameters or no matches are found, it should return an empty string.
$ curl localhost:8080/search
$ curl localhost:8080/search?q=mrs
Mrs. Rosalia Wisoky, (865) 611-8960
Mrs. Earl Gaylord, 944-345-3158
Mrs. Roslyn Moen, 526.643.3627
Mrs. Giovani Rempel, 842-246-9417


Theory:
In this lesson, we learned how to work with a request object and a response object (response) in Node.js .
The request object contains information about the incoming request, such as the URL, request method, headers, and request parameters. We can use the URL object to conveniently work with the URL and get the request parameters.
The response object is used to generate and send a response to a request. We can set the response code, headers and send the data in the response. We also discussed that the headers should be set before sending the response body.
We also looked at the events generated by the request and response objects. For example, the 'end' event is generated when the request is fully received, and we can process the request at this stage. The 'error' event is generated in case of an error and we can listen to it to process the error or register it.
As a result, we can create a server using the http.createServer() method and process incoming requests using the request and response objects. We can set the response code, headers and send the data in the response.
It is important to remember that in this lesson we have covered only the basics of working with request and response objects. In the future, we will explore more complex server scenarios in Node.js .