JS: HTTP Server 4 Routing
Implement the route /users/(\\w+).json, which will provide information about a specific entry from the directory. The data is given in json format, so you must set the correct media type.

$ curl localhost:4000/users/5.json
{"data":{"name":"Lonny McGlynn","phone":"(935) 384-0149"}}
If no such record exists, you must return the 404 response code and an empty body.

Theory:
This lesson discusses the concept of routing in web development. Routing is responsible for determining the handler for a specific requested page. The examples show static and dynamic routes. Static routes have an address that matches the route itself, while dynamic routes contain parameters. Regular expressions are used to define routes with parameters. In frameworks, routing is one of the main subsystems and provides a higher-level way to work with routes. It is important to consider the order of setting routes and handle errors such as "page not found". There are agreements on how to properly organize routing for CRUD operations (CREATE, READ, UPDATE, DELETE), taking into account the semantics of HTTP.