JS: HTTP Server 5 Automatic verification
server.js
Implement the creation of a new entry in the directory. The entry point will be the POST route to /users.json. Data:

{
  "name": "<name>",
  "phone": "<phone>"
}
On the server side, it is necessary to verify the correctness of the data (validate).

Request Example:

$ curl -XPOST \
>   -H 'Content-Type: application/json' \
>   --data '{"name":"Bob","phone":"912-114-23-22"}' \
>   localhost:4000/users.json
{"meta":{"location":"/users/10.json"},"data":{"name":"Bob","phone":"912-114-23-22","id":10}}
If successful, the response code 201 (created) and the body should be returned:

{
  "meta": {
    "location": "/users/10.json"
  },
  "data": {
"name": "Bob",
"phone": "912-114-23-22",
"id": 10
}
}
location - the address where you can get the data of this
id record - is generated using the nextId function and inserted into the record before being added to the general list
In case of validation errors, the response code 422 (unprocessable entity) and the body must be returned:

{
  "errors": [
    {
      "source": "name",
      "title" :"bad format"
    },
    {
      "source": "phone",
      "title" :"can't be blank"
    }
  ]
}
The number of errors and their type depends on the data received.

user.js
Implement a function to validate the data that came from the form according to the following rules:

The name cannot be empty
The name must match the pattern ^[\w\.]+$
The phone cannot be empty
The function should return an empty array if there are no errors and an array with errors if there are any. Each element of the array is an object of the following structure:

{
  source: <field-name>,
  title: <message>,
}
Types of messages:

If the field is empty: can't be blank
If the format does not match the desired one: bad format

Theory:
In this lesson, we learned how to write data to our phone book via telnet. To do this, we used the POST request method, which is not idempotent, that is, resending data can lead to duplicates or an error.

When sending the POST request, we specified two headers: Content-Type, so that the server knows how to parse our body, and Content-Length, to determine the end of the request.

On the server, we have created a handler for the POST request, which accepts data in chunks and responds to the request after full processing of the data.

We also discussed the importance of validating data to check it for correctness and compliance with format requirements. Validation helps to prevent possible vulnerabilities and errors in the system.

Depending on the situation, we have identified three possible responses to the POST request: successful request (API), successful request (Web) and validation error. Each of them has its own response code and response body format.

Bottom line: in this lesson, we learned how to write data to the phone book via telnet, process POST requests on the server, validate data and determine the correct format for responding to a request.