JS: Express 2 Assets
Implement an application that is a blog. The site should allow you to view the list of posts and allow you to add new posts.

GET /posts - list of posts
GET /posts/:id - page of the post
GET /posts/new - form for creating a new post
POST /posts - create a new post. If successful, a redirect to the post page is expected, otherwise code 422 should be returned in response to the request and the same form for creating a new post with errors should be drawn
The application contains one entity - Post, which contains three required fields:

title and body - set via
the id form - is put down automatically, for each new post it must increase by one
entities/Post
Implement the Post entity

solution.js
Implement the missing express handlers

views/posts/index.pug
Implement the output of posts in tabular form. There should also be links on the page to view posts and a link to create a new one.

views/posts/new.pug
Implement the form for creating a new post

views/posts/show.pug
Implement viewing a specific post

Theory:
This lesson discusses the problem of generating HTML code on the server when processing HTTP requests. It is shown that the manual generation of HTML code is inconvenient and has many disadvantages. Then two approaches to templating are presented: jinja-like and haml-like.

Jinja-like templating tools, such as Jinja in Python, suggest using special constructs to insert data into HTML code. However, such template engines have disadvantages, such as the difficulty of editing templates, problems with HTML semantics and dependence on the coding style.

Haml-like templating tools such as Pug (formerly known as Jade) in JavaScript offer a simpler and more compact syntax for describing HTML code. They do not allow you to write arbitrary code outside of control structures and provide template validation.

The lesson also discusses the integration of Pug with Express, a framework for creating web applications on Node.js . It shows how to install Pug as a template rendering engine and how to pass parameters to the template.

The lesson also discusses additional Pug features, such as template inheritance and inclusions.

Lesson summary: instead of manually generating HTML code on the server, it is recommended to use template engines such as Pug for more convenient and flexible creation of HTML pages. Pug offers a simple and compact syntax, provides template validation, and allows inheritance and inclusions to be used for code reuse.